backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot[all]==21.0.1
redis[hiredis]==5.0.0
chromadb>=0.4.24
sentence-transformers>=2.2.2
huggingface_hub>=0.16.4
openai>=1.30.0
apscheduler==3.10.4
pydantic==2.5.1
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
httpx>=0.27.0
sqlalchemy==2.0.23
asyncpg==0.29.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
torch==2.1.1
transformers==4.36.2
datasets==2.15.0
peft==0.7.1
accelerate==0.25.0